requirements.txt
absl==0.0
colorama==0.4.6
estimators==0.1.0.dev0
masking==0.0.3
matplotlib==3.10.1
numba==0.61.0
numpy==2.2.3
ordered_set==4.1.0
pandas==2.2.3
psycopg2==2.9.10
pynvml==12.0.0
pytorch_lamb==1.0.0
PyYAML==6.0.2
scipy==1.15.2
seaborn==0.13.2
setproctitle==1.3.5
setuptools==75.1.0
sqlglot==26.9.0
sympy==1.13.3
torch==2.6.0
tqdm==4.66.5
wandb==0.19.8
yagmail==0.15.293
